you-play-app/streaming-service/requirements.txt
minio>=7.2.0
boto3>=1.26.0
fastapi>=0.68.0
uvicorn>=0.15.0
python-dotenv>=0.19.0 